+++ InternetArchive.NET.Tests/WaybackTests.cs
@@ -31,16 +31,25 @@ public class WaybackTests
         Assert.AreEqual(25, response2.Results.Count);
         Assert.AreEqual(25, response2.Results.Select(x => x.Timestamp).Except(response1.Results.Select(x => x.Timestamp)).Count());
 
-        var response3 = await _client.Wayback.SearchAsync(
-            new Wayback.SearchRequest
-            {
-                Url = "www.experimentaltvcenter.org",
-                Limit = 25,
-                Offset = 10
-            }
-        );
+#pragma warning disable CS0612 // Type or member is obsolete
+        try
+        {
+            var response3 = await _client.Wayback.SearchAsync(
+                new Wayback.SearchRequest
+                {
+                    Url = "www.experimentaltvcenter.org",
+                    Limit = 25,
+                    Offset = 10
+                }
+            );
 
-        Assert.AreEqual(0, response3.Results.Select(x=> x.Timestamp).Except(response2.Results.Select(x=> x.Timestamp)).Count());
+            Assert.AreEqual(0, response3.Results.Select(x => x.Timestamp).Except(response2.Results.Select(x => x.Timestamp)).Count());
+        }
+        catch (InternetArchiveException ex)
+        {
+            Assert.IsTrue(ex.Message.Contains("no longer supported"));
+        }
+#pragma warning restore CS0612
     }
 
     [TestMethod]
@@ -82,56 +91,5 @@ public class WaybackTests
             if (response.ResumeKey == null) break;
             request.ResumeKey = response.ResumeKey;
         }
-    }
-
-    [TestMethod]
-    public async Task GetNumPagesAsync()
-    {
-        var pages = await _client.Wayback.GetNumPagesAsync("www.experimentaltvcenter.org");
-        Assert.IsNotNull(pages);
-    }
-
-    [TestMethod]
-    public async Task IsAvailableAsync()
-    {
-        var response = await _client.Wayback.IsAvailableAsync("www.experimentaltvcenter.org");
-
-        Assert.IsNotNull(response);
-        Assert.IsTrue(response.IsAvailable);
-        Assert.IsNotNull(response.Url);
-        Assert.IsNotNull(response.Timestamp);
-        Assert.AreEqual(HttpStatusCode.OK, response.Status);
-    }
-
-    [TestMethod]
-    public async Task IsAvailableTimestamp()
-    {
-        var response = await _client.Wayback.IsAvailableAsync("www.bombfactory.com", new DateTime(2000, 7, 4));
-
-        Assert.IsNotNull(response);
-        Assert.IsTrue(response.IsAvailable);
-        Assert.IsNotNull(response.Url);
-        Assert.AreEqual(2000, response.Timestamp?.Year);      
-    }
-
-    [TestMethod]
-    public async Task IsAvailableTimestampAsync()
-    {
-        var response = await _client.Wayback.IsAvailableAsync("www.experimentaltvcenter.org", new DateTime(2001, 3, 31));
-
-        Assert.IsNotNull(response);
-        Assert.IsTrue(response.IsAvailable);
-        Assert.IsNotNull(response.Url);
-        Assert.AreEqual(2001, response.Timestamp?.Year);
-        Assert.AreEqual(HttpStatusCode.OK, response.Status);
-    }
-
-    [TestMethod]
-    public async Task IsNotAvailableAsync()
-    {
-        var response = await _client.Wayback.IsAvailableAsync("www.experimentaltvcenter.org__");
-
-        Assert.IsNotNull(response);
-        Assert.IsFalse(response.IsAvailable);
     }
 }